Separate the responsability of the process
