+++ DeveloperJava.java
@@ -12,10 +12,6 @@ public class DeveloperJava implements Developer {
 
     private Boolean isIssueAnalyzed = Boolean.FALSE;
 
-    private Boolean isSolutionCommited = Boolean.FALSE;
-
-    // Verificar possibilidade de aplicas state
-
     public DeveloperJava(String name, String issue) {
         this.name = name;
         this.issue = issue;
@@ -27,44 +23,22 @@ public class DeveloperJava implements Developer {
 
         out.println(String.format("Solução desenvolvida em Java pelo [%s]", this.name));
 
-        try {
-            commitSolution();
-            this.isDevelopmentFinished = TRUE;
-        } catch (MergeProblemsException e) {
-            resolveConflictAndCommit();
-        } catch (Exception ex) {
-            err.println("Not expected !! Nothing was commited");
-            err.println(String.format("[%s]: Sabia não!", name));
-        }
-
-        if (isAllProcessDone()) {
-            out.println(String.format("Parabéns [%s] pelo trabalho!", this.name));
-        }
+        isDevelopmentFinished = TRUE;
     }
 
     @Override
-    public void commitSolution() throws MergeProblemsException {
-        throw new MergeProblemsException();
-//        throw new NullPointerException();
-//        out.println("Solução commitada no git");
-//        this.isSolutionCommited = TRUE;
-//        this.isDevelopmentFinished = TRUE;
-    }
-
-    private void analyzeIssue(String issue) {
+    public void analyzeIssue(String issue) {
         out.println(String.format("Issue analisada [%s]", issue));
         this.isIssueAnalyzed = TRUE;
     }
 
-    private void resolveConflictAndCommit(){
-        out.println("Conflito resolvido,  código mergeado e commitado");
-        this.isSolutionCommited = TRUE;
-        this.isDevelopmentFinished = TRUE;
+    @Override
+    public Boolean isDevelopmentFinished() {
+        return this.isDevelopmentFinished;
     }
 
-    public Boolean isAllProcessDone() {
-        return isIssueAnalyzed &&
-                isSolutionCommited &&
-                isDevelopmentFinished;
+    @Override
+    public String getName() {
+        return name;
     }
 }
