+++ Processo.java
@@ -2,8 +2,11 @@ public class Processo {
 
     public static void main(String[] args) {
         Developer dev = new DeveloperJava("Marinho", "VP-99999");
-
         dev.developmentOfSolution();
+
+        CodeCommiter codeCommiter = new CodeCommiter(dev);
+
+        codeCommiter.commitCode();
     }
 
 }
